fixup! Fix merge: rv32i state integer size widths
This applies essentially the same fixes as
57b630b2dcf51932f890f452ac89df92b68bbb86, however the local CI did
only test the CI's mock function implementations and did not compile
for the real architecture.

+++ arch/rv32i/src/syscall.rs
@@ -367,7 +367,7 @@ impl kernel::syscall::UserspaceKernelBoundary for SysCall {
           : "memory"
           : "volatile");
 
-        let ret = match mcause::Trap::from(state.mcause) {
+        let ret = match mcause::Trap::from(state.mcause as usize) {
             mcause::Trap::Interrupt(_intr) => {
                 // An interrupt occurred while the app was running.
                 ContextSwitchReason::Interrupted
